requirements.txt
joblib==1.5.1
numpy==2.3.0
pillow==11.2.1
scikit-learn==1.7.0
scipy==1.15.3
threadpoolctl==3.6.0
